- Se limpió el código sacando los tipos a un archivo env.d.ts, se pasó la función customError a una carpeta separada con el nombre createCustomError y se pasó el archivo config a una carpeta config.
