+++ src/private-api.ts
@@ -1,33 +1,8 @@
 import axios from "axios";
-import config from "./config";
-
-const customError = (code: number, message: string): Error => {
-    const error = new Error();
-    return Object.assign(error, { code, message: `[${config.lib.name}] ${message}` });
-};
-
-interface PhoneData {
-    iso: any;
-    phone: string;
-};
-
-interface CreditCardData {
-    pan: string | number;
-    expirationDate?: string;
-    cvc?: string | number;
-    cvv?: string | number;
-};
-
-interface Data {
-    email?: string;
-    phone?: PhoneData;
-    domain?: string;
-    creditCard?: string | CreditCardData;
-    ip?: string;
-};
+import { createCustomError } from "./errors/custom-error";
 
 export const isValidData = async (token: string | null, data: Data): Promise<any> => {
-    if (token === null) throw customError(3000, "Invalid private token.");
+    if (token === null) throw createCustomError(3000, "Invalid private token.");
     
     let i = false;
     for (const key in data) {
@@ -36,7 +11,7 @@ export const isValidData = async (token: string | null, data: Data): Promise<any
             break;
         }
     }
-    if (!i) throw customError(1500, "You must provide at least one parameter.");
+    if (!i) throw createCustomError(1500, "You must provide at least one parameter.");
     
     try {
         const response = await axios.post("https://api.tpeoficial.com/v1/private/secure/verify", data, {
@@ -44,6 +19,6 @@ export const isValidData = async (token: string | null, data: Data): Promise<any
         });
         return response.data;
     } catch (error: any) {
-        throw customError(5000, error.message);
+        throw createCustomError(5000, error.message);
     }
 };
